Make package more global
I used a custom version of this package a lot but never put it in git. So decided to rewrite this entire package now to be more global and not only for images but so it can be used by any file type. This is the way I used a lot for storage configuration without having to know where the real file is stored (can be database, disk, api etc) it's up to you how you want to implement it.